+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "react-portfolio",
+  "name": "house-love",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
